@@ -1,97 +1,103 @@
 {
-  "name": "appaydin/pd-admin",
-  "description": "Symfony Powerful Admin Dashboard",
-  "homepage": "https://github.com/appaydin/pd-admin",
-  "type": "project",
-  "license": "MIT",
-  "authors": [
-    {
-      "name": "Ramazan APAYDIN",
-      "homepage": "http://github.com/appaydin"
-    }
-  ],
-  "require": {
-    "php": "^7.3.0",
-    "ext-ctype": "*",
-    "ext-iconv": "*",
-    "ext-json": "*",
-    "appaydin/pd-mailer": "2.*",
-    "appaydin/pd-menu": "2.*",
-    "appaydin/pd-user": "2.*",
-    "appaydin/pd-widget": "2.*",
-    "beberlei/doctrineextensions": "1.*",
-    "intervention/image": "2.*",
-    "knplabs/knp-paginator-bundle": "5.*",
-    "sensio/framework-extra-bundle": "5.*",
-    "symfony/asset": "5.*",
-    "symfony/console": "5.*",
-    "symfony/dotenv": "5.*",
-    "symfony/expression-language": "5.*",
-    "symfony/flex": "1.*",
-    "symfony/form": "5.*",
-    "symfony/framework-bundle": "5.*",
-    "symfony/mailer": "5.*",
-    "symfony/monolog-bundle": "3.*",
-    "symfony/orm-pack": "1.*",
-    "symfony/process": "5.*",
-    "symfony/security-bundle": "5.*",
-    "symfony/translation": "5.*",
-    "symfony/twig-bundle": "5.*",
-    "symfony/validator": "5.*",
-    "symfony/yaml": "5.*"
-  },
-  "require-dev": {
-    "roave/security-advisories": "dev-master",
-    "symfony/debug-bundle": "5.*",
-    "symfony/maker-bundle": "^1.0",
-    "symfony/profiler-pack": "^1.0",
-    "symfony/test-pack": "^1.0"
-  },
-  "config": {
-    "preferred-install": {
-      "*": "dist"
-    },
-    "sort-packages": true
-  },
-  "autoload": {
-    "psr-4": {
-      "App\\": "src/"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "App\\Tests\\": "tests/"
-    }
-  },
-  "replace": {
-    "paragonie/random_compat": "2.*",
-    "symfony/polyfill-ctype": "*",
-    "symfony/polyfill-iconv": "*",
-    "symfony/polyfill-php73": "*",
-    "symfony/polyfill-php72": "*",
-    "symfony/polyfill-php71": "*",
-    "symfony/polyfill-php70": "*",
-    "symfony/polyfill-php56": "*"
-  },
-  "scripts": {
-    "auto-scripts": {
-      "cache:clear": "symfony-cmd",
-      "assets:install %PUBLIC_DIR%": "symfony-cmd"
-    },
-    "post-install-cmd": [
-      "@auto-scripts"
+    "name": "appaydin/pd-admin",
+    "description": "Symfony Powerful Admin Dashboard",
+    "homepage": "https://github.com/appaydin/pd-admin",
+    "type": "project",
+    "license": "MIT",
+    "authors": [
+        {
+            "name": "Ramazan APAYDIN",
+            "homepage": "http://github.com/appaydin"
+        }
     ],
-    "post-update-cmd": [
-      "@auto-scripts"
-    ]
-  },
-  "conflict": {
-    "symfony/symfony": "*"
-  },
-  "extra": {
-    "symfony": {
-      "allow-contrib": false,
-      "require": "5.0.*"
+    "require": {
+        "php": "^7.4.0",
+        "ext-ctype": "*",
+        "ext-iconv": "*",
+        "ext-json": "*",
+        "ext-mbstring": "*",
+        "appaydin/pd-mailer": "2.*",
+        "appaydin/pd-menu": "3.*",
+        "appaydin/pd-user": "3.*",
+        "appaydin/pd-widget": "^2.0",
+        "beberlei/doctrineextensions": "1.*",
+        "composer/package-versions-deprecated": "1.11.99.1",
+        "doctrine/doctrine-bundle": "^2.2",
+        "doctrine/doctrine-migrations-bundle": "^3.0",
+        "doctrine/orm": "^2.8",
+        "intervention/image": "2.*",
+        "knplabs/knp-paginator-bundle": "5.*",
+        "sensio/framework-extra-bundle": "5.*",
+        "symfony/asset": "5.*",
+        "symfony/console": "5.*",
+        "symfony/dotenv": "5.*",
+        "symfony/expression-language": "5.*",
+        "symfony/flex": "1.*",
+        "symfony/form": "5.*",
+        "symfony/framework-bundle": "5.*",
+        "symfony/mailer": "5.*",
+        "symfony/monolog-bundle": "3.*",
+        "symfony/proxy-manager-bridge": "5.2.*",
+        "symfony/security-bundle": "5.*",
+        "symfony/translation": "5.*",
+        "symfony/twig-bundle": "5.*",
+        "symfony/validator": "5.*",
+        "symfony/webpack-encore-bundle": "^1.8",
+        "symfony/yaml": "5.*",
+    },
+    "require-dev": {
+        "roave/security-advisories": "dev-master",
+        "symfony/debug-bundle": "5.*",
+        "symfony/maker-bundle": "^1.0",
+        "symfony/profiler-pack": "^1.0",
+        "symfony/test-pack": "^1.0"
+    },
+    "config": {
+        "optimize-autoloader": true,
+        "preferred-install": {
+            "*": "dist"
+        },
+        "sort-packages": true
+    },
+    "autoload": {
+        "psr-4": {
+            "App\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "App\\Tests\\": "tests/"
+        }
+    },
+    "replace": {
+        "symfony/polyfill-ctype": "*",
+        "symfony/polyfill-iconv": "*",
+        "symfony/polyfill-php73": "*",
+        "symfony/polyfill-php72": "*",
+        "symfony/polyfill-php71": "*",
+        "symfony/polyfill-php70": "*",
+        "symfony/polyfill-php56": "*",
+        "symfony/polyfill-mbstring": "*"
+    },
+    "scripts": {
+        "auto-scripts": {
+            "cache:clear": "symfony-cmd",
+            "assets:install %PUBLIC_DIR%": "symfony-cmd"
+        },
+        "post-install-cmd": [
+            "@auto-scripts"
+        ],
+        "post-update-cmd": [
+            "@auto-scripts"
+        ]
+    },
+    "conflict": {
+        "symfony/symfony": "*"
+    },
+    "extra": {
+        "symfony": {
+            "allow-contrib": false,
+            "require": "5.2.*"
+        }
     }
-  }
 }
