@@ -19,7 +19,7 @@ use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
  * User Accounts.
  *
  * @ORM\Table(name="user")
- * @ORM\Entity(repositoryClass="App\Repository\UserRepository")
+ * @ORM\Entity(repositoryClass="App\Repository\Account\UserRepository")
  * @UniqueEntity(fields="email", message="email_already_taken")
  *
  * @author Ramazan APAYDIN <apaydin541@gmail.com>
